# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPay API",
        "description": "Documentation de l'API OMPay (Orange Money Pay)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local"
        },
        {
            "url": "https://ompay.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/auth/connexion": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur sans code PIN (compte déjà activé)",
                "operationId": "e76ae49b81740147572539cd72ad43c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "password"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771428150"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion effectué avec succès!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOi..."
                                        },
                                        "refresh_token": {
                                            "type": "string",
                                            "example": "eyJhbGciOi..."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion effectué avec succès!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    },
                    "403": {
                        "description": "Compte non activé"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription utilisateur (client/distributeur)",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "telephone",
                                    "password",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Fall"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Awa"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "client",
                                            "distributeur"
                                        ],
                                        "example": "client"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
...
}